feat: centralize more collisions
