+++ frontend/src/main.js
@@ -1,12 +1,15 @@
 import {
   spawnPlayer,
-  circleRectColliding,
   SEGMENT_SIZE,
   PLAYER_MAX_HEALTH,
   createSpawnDoor,
   spawnContainers,
   openContainer,
 } from "./game_logic.js";
+import {
+  circleRectColliding,
+  checkAllCollisions,
+} from "./systems/collision-system.js";
 import { moveTowards, isColliding } from "./utils/geometry.js";
 import {
   moveZombie,
@@ -47,20 +50,10 @@ import {
 } from "./inventory.js";
 import { RECIPES, canCraft, craftRecipe } from "./crafting.js";
 import { dropLoot } from "./loot.js";
-import {
-  createFireball,
-  updateFireballs,
-  predictFireballEndpoint,
-  fireballStats,
-  updateExplosions,
-} from "./spells.js";
-import {
-  createArrow,
-  updateArrows,
-  predictArrowEndpoint,
-} from "./entities/arrow.js";
 import { SKILL_INFO, SKILL_UPGRADERS } from "./skill_tree.js";
-import { createOrbs, updateOrbs } from "./entities/orbs.js";
+import { createOrbs } from "./entities/orbs.js";
+import { updateAbilities } from "./systems/ability-system.js";
+import { render as renderScene } from "./systems/rendering-system.js";
 import { makeDraggable } from "./ui.js";
 import { createInventoryUI } from "./components/inventory-ui.js";
 import { createSkillTreeUI } from "./components/skill-tree-ui.js";
@@ -241,7 +234,6 @@ function renderHotbar() {
     getItemCooldown,
   );
 }
-
 
 function toggleInventory(open) {
   if (open === inventoryOpen) return;
@@ -587,66 +579,7 @@ function update() {
     }
   }
 
-  for (let i = worldItems.length - 1; i >= 0; i--) {
-    const it = worldItems[i];
-    if (isColliding(player, it, 10)) {
-      if (addItem(inventory, it.type, it.count)) {
-        worldItems.splice(i, 1);
-        hud.showPickupMessage(`Picked up ${it.type}`);
-        renderInventory();
-        renderHotbar();
-      }
-    }
-  }
-
   if (player.swingTimer > 0) player.swingTimer--;
-
-  if (fireballCooldown > 0) fireballCooldown--;
-
-  if (
-    activeSlot &&
-    activeSlot.item === "fireball_spell" &&
-    player.abilities.fireball &&
-    useHeld &&
-    fireballCooldown <= 0
-  ) {
-    if (countItem(inventory, "fire_core") > 0) {
-      removeItem(inventory, "fire_core", 1);
-      const dir = { x: mousePos.x - player.x, y: mousePos.y - player.y };
-      const fb = createFireball(
-        player.x,
-        player.y,
-        dir,
-        player.abilities.fireballLevel,
-        player.damageBuffMult,
-      );
-      if (fb) fireballs.push(fb);
-      fireballCooldown = 15;
-      renderInventory();
-      renderHotbar();
-    } else {
-      hud.showPickupMessage("Out of Fire Cores!");
-    }
-  }
-
-  if (player.weapon && player.weapon.type === "bow") {
-    bowAiming = aimHeld;
-    if (aimRelease) {
-      const arrowsLeft = countItem(inventory, "arrow");
-      if (arrowsLeft > 0) {
-        removeItem(inventory, "arrow", 1);
-        const dir = { x: mousePos.x - player.x, y: mousePos.y - player.y };
-        const a = createArrow(player.x, player.y, dir, player.damageBuffMult);
-        if (a) arrows.push(a);
-        renderInventory();
-        renderHotbar();
-      } else {
-        hud.showPickupMessage("Out of Arrows!");
-      }
-    }
-  } else {
-    bowAiming = false;
-  }
 
   if (player.weapon && useHeld && player.swingTimer <= 0) {
     const killed = attackZombiesWithKills(
@@ -708,29 +641,41 @@ function update() {
     }
   }
 
-  if (player.abilities.fireOrb) {
-    updateOrbs(fireOrbs, player, zombies, player.abilities.fireOrbLevel, (z) =>
-      dropLoot(z, worldItems),
-    );
-  }
-
-  updateFireballs(fireballs, zombies, walls, explosions, (z) =>
-    dropLoot(z, worldItems),
-  );
-  updateExplosions(explosions);
-  updateArrows(
+  const abilityState = updateAbilities({
+    player,
+    inventory,
+    fireballs,
+    fireOrbs,
     arrows,
     zombies,
+    mousePos,
+    worldItems,
+    hud,
+    renderInventory,
+    renderHotbar,
+    useHeld,
+    aimHeld,
+    aimRelease,
+    fireballCooldown: { value: fireballCooldown },
+  });
+  bowAiming = abilityState.bowAiming;
+
+  checkAllCollisions({
+    player,
+    zombies,
+    arrows,
+    fireballs,
     walls,
-    (z) => dropLoot(z, worldItems),
-    (w) =>
-      worldItems.push({
-        x: w.x + SEGMENT_SIZE / 2,
-        y: w.y + SEGMENT_SIZE / 2,
-        type: MATERIAL_DROPS[w.material],
-        count: 1,
-      }),
-  );
+    explosions,
+    worldItems,
+    inventory,
+    hud,
+    renderInventory,
+    renderHotbar,
+    dropLoot,
+    materialDrops: MATERIAL_DROPS,
+  });
+  fireballCooldown = abilityState.fireballCooldown;
   hud.update();
 
   if (!victory && zombies.length === 0) {
@@ -744,158 +689,32 @@ function update() {
   renderHotbar();
 }
 
-function render() {
-  ctx.clearRect(0, 0, canvas.width, canvas.height);
-
-  walls.forEach((w) => {
-    const img = WALL_IMAGES[w.material];
-    if (img && img.complete) {
-      ctx.globalAlpha = w.opened ? 0.5 : 1;
-      ctx.drawImage(img, w.x, w.y, SEGMENT_SIZE, SEGMENT_SIZE);
-      ctx.globalAlpha = 1;
-    } else {
-      ctx.fillStyle = "gray";
-      ctx.fillRect(w.x, w.y, SEGMENT_SIZE, SEGMENT_SIZE);
-    }
-    if (w.damageTimer > 0) {
-      ctx.fillStyle = "rgba(255,0,0,0.5)";
-      ctx.fillRect(w.x, w.y, SEGMENT_SIZE, SEGMENT_SIZE);
-    }
-    if (w.hp < w.maxHp) {
-      ctx.fillStyle = "red";
-      ctx.fillRect(w.x, w.y - 6, SEGMENT_SIZE, 4);
-      ctx.fillStyle = "lime";
-      ctx.fillRect(w.x, w.y - 6, (w.hp / w.maxHp) * SEGMENT_SIZE, 4);
-    }
-  });
-  containers.forEach((c) => {
-    ctx.globalAlpha = c.opened ? 0.5 : 1;
-    ctx.drawImage(cardboardBoxImg, c.x - 10, c.y - 10, 20, 20);
-    ctx.globalAlpha = 1;
-  });
-  if (spawnDoor) {
-    ctx.fillStyle = "brown";
-    ctx.fillRect(spawnDoor.x - 5, spawnDoor.y - 5, 10, 10);
-  }
-
-  drawSprite(ctx, playerSprite, player.x, player.y, player.facing);
-
-  if (player.abilities.fireOrb) {
-    ctx.fillStyle = "orange";
-    fireOrbs.forEach((o) => {
-      if (o.cooldown <= 0) {
-        ctx.beginPath();
-        ctx.arc(o.x, o.y, 5, 0, Math.PI * 2);
-        ctx.fill();
-      }
-    });
-  }
-
-  if (player.swingTimer > 0) {
-    ctx.strokeStyle = "orange";
-    ctx.lineWidth = 3;
-    const startA = Math.atan2(player.facing.y, player.facing.x) - Math.PI / 4;
-    ctx.beginPath();
-    ctx.arc(player.x, player.y, 25, startA, startA + Math.PI / 2);
-    ctx.stroke();
-    ctx.lineWidth = 1;
-  }
-  hud.render(ctx, player, inventory, countItem);
-
-  if (weapon) {
-    const img = ITEM_IMAGES[weapon.type];
-    if (img) {
-      ctx.drawImage(img, weapon.x - 8, weapon.y - 8, 16, 16);
-    } else {
-      ctx.fillStyle = "orange";
-      ctx.beginPath();
-      ctx.arc(weapon.x, weapon.y, 6, 0, Math.PI * 2);
-      ctx.fill();
-    }
-  }
-
-  ctx.fillStyle = "yellow";
-  worldItems.forEach((it) => {
-    ctx.beginPath();
-    ctx.arc(it.x, it.y, 5, 0, Math.PI * 2);
-    ctx.fill();
-  });
-
-  ctx.fillStyle = "brown";
-  arrows.forEach((a) => {
-    ctx.beginPath();
-    ctx.arc(a.x, a.y, 2, 0, Math.PI * 2);
-    ctx.fill();
-  });
-  const aimingBow =
-    player.weapon && player.weapon.type === "bow" && (keys[" "] || keys.mouse2);
-  if (aimingBow) {
-    const dir = { x: mousePos.x - player.x, y: mousePos.y - player.y };
-    const end = predictArrowEndpoint(player.x, player.y, dir, walls, zombies);
-    ctx.strokeStyle = "rgba(255,0,0,0.6)";
-    ctx.setLineDash([5, 5]);
-    ctx.beginPath();
-    ctx.moveTo(player.x, player.y);
-    ctx.lineTo(end.x, end.y);
-    ctx.stroke();
-    ctx.setLineDash([]);
-    ctx.beginPath();
-    ctx.arc(end.x, end.y, 3, 0, Math.PI * 2);
-    ctx.stroke();
-  }
-
-  const activeSlot = getActiveHotbarItem(inventory);
-  if (
-    activeSlot &&
-    activeSlot.item === "fireball_spell" &&
-    player.abilities.fireball
-  ) {
-    const dir = { x: mousePos.x - player.x, y: mousePos.y - player.y };
-    const end = predictFireballEndpoint(
-      player.x,
-      player.y,
-      dir,
-      walls,
-      zombies,
-    );
-    ctx.strokeStyle = "rgba(255,0,0,0.6)";
-    ctx.setLineDash([5, 5]);
-    ctx.beginPath();
-    ctx.moveTo(player.x, player.y);
-    ctx.lineTo(end.x, end.y);
-    ctx.stroke();
-    ctx.setLineDash([2, 4]);
-    const { radius } = fireballStats(player.abilities.fireballLevel);
-    ctx.beginPath();
-    ctx.arc(end.x, end.y, radius, 0, Math.PI * 2);
-    ctx.stroke();
-    ctx.setLineDash([]);
-  }
-
-  ctx.fillStyle = "orange";
-  fireballs.forEach((fb) => {
-    ctx.beginPath();
-    ctx.arc(fb.x, fb.y, 4, 0, Math.PI * 2);
-    ctx.fill();
-  });
-
-  ctx.fillStyle = "rgba(255,0,0,0.5)";
-  explosions.forEach((ex) => {
-    ctx.beginPath();
-    ctx.arc(ex.x, ex.y, ex.radius, 0, Math.PI * 2);
-    ctx.fill();
-  });
-
-  renderZombies(ctx, zombies, zombieSprite, fireZombieSprite);
-
-  // The overlay div already shows the Game Over message
-  // and restart button, so nothing is drawn here when
-  // the player loses.
-}
-
 function gameLoop() {
   update();
-  render();
+  renderScene(ctx, {
+    walls,
+    containers,
+    spawnDoor,
+    player,
+    playerSprite,
+    fireZombieSprite,
+    zombieSprite,
+    fireOrbs,
+    hud,
+    inventory,
+    ITEM_IMAGES,
+    weapon,
+    worldItems,
+    arrows,
+    mousePos,
+    zombies,
+    activeSlot: getActiveHotbarItem(inventory),
+    fireballs,
+    explosions,
+    bowAiming,
+    cardboardBoxImg,
+    countItem,
+  });
   requestAnimationFrame(gameLoop);
 }
